% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbssuitabilityhaiti-package.R
\docType{package}
\name{cbssuitabilityhaiti-package}
\alias{cbssuitabilityhaiti}
\alias{cbssuitabilityhaiti-package}
\title{cbssuitabilityhaiti: The Package contains data from the Soil projects in Haiti}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sebastian Camilo Loos \email{seba.loos@hotmail.com} (0000-0002-8830-1734)

Authors:
\itemize{
  \item Max Power \email{email@email.com} (0000-0002-8830-1734)
}

}
\keyword{internal}
