% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwater.R
\docType{data}
\name{mwater}
\alias{mwater}
\title{Location of different sanitation systems in Haiti}
\format{
A tibble with 1849 rows and 7 variables:
\describe{
\item{Latitude.o}{NA}
\item{Longitude}{NA}
\item{Administra}{NA}
\item{Type}{NA}
\item{Date.added}{NA}
\item{Datasets..}{NA}
\item{geometry}{NA}
}
}
\usage{
mwater
}
\description{
Description
Description
Description
}
\keyword{datasets}
