% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwater.R
\docType{data}
\name{mwater}
\alias{mwater}
\title{Location of different sanitation systems in Haiti}
\format{
A tibble with 1849 rows and 7 variables:
\describe{
\item{Latitude.o}{Lattitude coordinate}
\item{Longitude}{Longitude coordinate}
\item{Administra}{Communal section- smallest administrative unit in Haiti}
\item{Type}{Type of water access point}
\item{Date.added}{Date water access point was added to mWater}
\item{Datasets..}{Dataset in mWater that point is part of, including organizaiton that is responsible for data}
\item{geometry}{Geospatial data of the different access points that were added to mWater}
}
}
\usage{
mwater
}
\description{
Description
Description
Description
}
\keyword{datasets}
