% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwater.R
\docType{data}
\name{mwater}
\alias{mwater}
\title{Location of different sanitation systems in Haiti}
\format{
A tibble with 1849 rows and 7 variables:
\describe{
\item{latitude}{Lattitude coordinate}
\item{longitude}{Longitude coordinate}
\item{administra}{Communal section- smallest administrative unit in Haiti}
\item{type}{Type of water access point}
\item{date_added}{Date water access point was added to mWater}
\item{datasets}{Dataset in mWater that point is part of, including organizaiton that is responsible for data}
\item{geometry}{Geospatial data of the different access points that were added to mWater}
}
}
\usage{
mwater
}
\description{
This data includes data from a ........ done for the city
of Cap Haitien, Haiti in 2022. The data collection and characterization was
done between 2016 and 2022.
}
\keyword{datasets}
