% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okap.R
\docType{data}
\name{okap}
\alias{okap}
\title{Categorized data for neighborhoods in Haiti}
\format{
A tibble with 198 rows and 21 variables:
#' \describe{
\item{neighborho}{NA}
\item{name}{NA}
\item{sup_km2}{NA}
\item{sector_id}{NA}
\item{cte}{NA}
\item{qlty_water}{NA}
\item{qty_water}{NA}
\item{health_car}{NA}
\item{schooling}{NA}
\item{transport}{NA}
\item{economy}{NA}
\item{ranking}{NA}
\item{sup_bati}{NA}
\item{density}{NA}
\item{aptitude}{NA}
\item{zoning}{NA}
\item{latrine}{NA}
\item{standing}{NA}
\item{density_ra}{NA}
\item{economy_nu}{NA}
\item{geometry}{NA}
}
}
\usage{
okap
}
\description{
Description
Description
Description
}
\keyword{datasets}
